% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analysoitu_otanta}
\alias{analysoitu_otanta}
\title{Analysoitu data satunnaisotannan jälkeen.}
\format{A data frame with the following structure:
\describe{
  \item{textid}{Tekstin id}
  \item{previous_paragraph}{Edellisen kappaleen teksti}
  \item{sentence}{Virke, jossa indikaattori on}
  ...
}}
\usage{
analysoitu_otanta
}
\description{
Tässä dataframessa ovat tapaukset, jotka on jo ryhmitelty satunnaisotannan perusteella tehdyn laadullisen analyysin tuloksena.
}
\keyword{datasets}
