% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chaptertitles}
\alias{chaptertitles}
\title{Ensimmäistä asumista käsittelevää kappaletta edeltävä välitosikko ja tekstin id}
\format{A data frame with the following structure:
\describe{
  \item{textid}{Tekstin id}
  \item{chaptertitle}{Edellisen kappaleen väliotsikko}
  ...
}}
\usage{
chaptertitles
}
\description{
Tällä vain täydennetään jo olemassaolevaa dataa
}
\keyword{datasets}
