% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiying_words.R
\name{ClassifyVerbs}
\alias{ClassifyVerbs}
\title{Group verbs according to predefined categories
This is just a helper function}
\usage{
ClassifyVerbs(mydf)
}
\arguments{
\item{mydf}{a data frame}
}
\description{
Group verbs according to predefined categories
This is just a helper function
}
