% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{EditNames}
\alias{EditNames}
\title{muokkaa nopeasti nimiä}
\usage{
EditNames(oldnames, namedf)
}
\arguments{
\item{oldnames}{muokattava vektori}

\item{namedf}{data frame, jossa vanhat nimet (old) ja uudet (new)}
}
\description{
muokkaa nopeasti nimiä
}
