% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fn}
\alias{fn}
\title{Convinience function for formatting numbers}
\usage{
fn(n, numbers = 2)
}
\arguments{
\item{n}{the number to print}

\item{numbers}{how many numbers after comma when rounding (in case of decimal numbers)}
}
\value{
a string representing the number in a formatted name
}
\description{
Convinience function for formatting numbers
}
