% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_groups.R
\name{FormatForStatisticalAnalysis}
\alias{FormatForStatisticalAnalysis}
\title{Muotoilee analysoitua dataa niin, että sitä on helppo käyttää tilastollisessa analyysissa}
\usage{
FormatForStatisticalAnalysis(numericalgroups = F, simplenames = F)
}
\arguments{
\item{numericalgroups}{käytetäänkö ryhmien niminä numeroita}
}
\description{
Muotoilee analysoitua dataa niin, että sitä on helppo käyttää tilastollisessa analyysissa
}
