% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_groups.R
\name{GetGroupsFromYaml}
\alias{GetGroupsFromYaml}
\title{Hae data ryhmistä yaml-tiedostojen perusteella}
\usage{
GetGroupsFromYaml(otanta)
}
\description{
Hae data ryhmistä yaml-tiedostojen perusteella
}
