% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetInfo}
\alias{GetInfo}
\title{Print some essential information  about an example based on text id}
\usage{
GetInfo(textid)
}
\arguments{
\item{textid}{id of the text}
}
\description{
Print some essential information  about an example based on text id
}
