% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprels.R
\name{GetListOfDepRels}
\alias{GetListOfDepRels}
\title{List the different deprels as separate data frames}
\usage{
GetListOfDepRels()
}
\value{
the list with the data frames
}
\description{
List the different deprels as separate data frames
}
