% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups.meta}
\alias{groups.meta}
\title{Metatiedot ryhmist채}
\format{An object of class \code{data.frame} with 16 rows and 3 columns.}
\usage{
groups.meta
}
\description{
T채ss채 dataframessa ovat tarkemmat tiedot eri ryhmist채
}
\keyword{datasets}
