% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MoreInfo}
\alias{MoreInfo}
\title{Antaa nopeasti tietoa jonkin klusterin ominaisuuksista}
\usage{
MoreInfo(textid)
}
\arguments{
\item{textid}{esimerkin teksti-id}
}
\description{
Antaa nopeasti tietoa jonkin klusterin ominaisuuksista
}
