% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Output}
\alias{Output}
\title{Just a convinience function for outputting to the right folder}
\usage{
Output(odir = "output", fname = "koko_artikkeli.Rmd")
}
\arguments{
\item{odir}{the folder to output to, defaults to "output"}

\item{fname}{name of the file to output (defaults to "koko_artikkeli.Rmd")}
}
\description{
Just a convinience function for outputting to the right folder
}
