% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overall_stats}
\alias{overall_stats}
\title{Pari yleistilastoa koko aineistosta}
\format{An object of class \code{list} of length 7.}
\usage{
overall_stats
}
\description{
Ennen kaikkea kaikkien tekstien määrä (niidenkin, joissa ei asumiskpl)
}
\keyword{datasets}
