% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{previous_pars}
\alias{previous_pars}
\title{Ensimmäistä asumista käsittelevää kappaletta edeltävä kappale ja tekstin id}
\format{A data frame with the following structure:
\describe{
  \item{textid}{Tekstin id}
  \item{previous_paragraph}{Edellisen kappaleen teksti}
  ...
}}
\usage{
previous_pars
}
\description{
Tällä vain täydennetään jo olemassaolevaa dataa
}
\keyword{datasets}
