% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencing.R
\name{Ref}
\alias{Ref}
\title{Referencing tables, figures etc in the document.}
\usage{
Ref(reftype = character(), label = character(), caption = character())
}
\arguments{
\item{reftype}{what we are referencing e.g. "tab", "fig"}

\item{label}{identifier for this particular instance}

\item{caption}{the caption (omit when only referencing)}
}
\value{
the reference for the table, figure etc OR the caption, depending on the input parameters
}
\description{
Referencing tables, figures etc in the document.
}
\examples{

#1. Creating a caption for a figure
Ref('fig','ff_flower', 'A diagram of flowers')

#2. Creating a caption for a table
Ref('fig','tt_letters', 'A table of letters in text A')

Notice that a good practice is to start labels of one type
with a prefix like tt_ for tables. This makes it easier to use
autocomplete functions in text editors. The prefixes are not, however,
compulsory, Ref('table','just.a.label','A table of letters in text A') will
do just as well.

 #3. Referencing:
Look at figure Ref('fig','tt_letters'), which...

}
