% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ristiin_annotointi}
\alias{ristiin_annotointi}
\title{Ristiinannotoinnin tulokset kappaleittain}
\format{An object of class \code{data.frame} with 322 rows and 5 columns.}
\usage{
ristiin_annotointi
}
\description{
T채ss채 dataframessa ovat ne 322 kappaletta, jotka kirjoittajat molemmat tahoillaan annotoivat siten, ett채 kumpikin merkitsi joka kappaleeseen sen aiheen, jota kappale kirjoittajan mielest채 edusti.
}
\keyword{datasets}
