% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiying_words.R
\name{VerbInfo}
\alias{VerbInfo}
\title{Classify some verbs}
\usage{
VerbInfo(mylist)
}
\arguments{
\item{mylist}{a list containing a data frame}
}
\value{
the list containing a sublist of verbs
}
\description{
Classify some verbs
}
